# backend/Entities/Order.cs
using System.ComponentModel.DataAnnotations.Schema;
using System.ComponentModel.DataAnnotations;

namespace backend.Entities
{
    public enum OrderStatus
    {
        Pending,
        Processing,
        Shipped,
        Delivered,
        Cancelled,
    }

    public class Order
    {
        [Key]
        public int OrderID { get; set; }

        [Required]
        public DateTime DateTime { get; set; }

        [Required]
        [Column(TypeName = "decimal(18,2)")]
        public decimal TotalPrice { get; set; }

        [Required]
        public OrderStatus Status { get; set; }
        public string PaymentType {  get; set; }

        [Required]
        public string TransactionID { get; set; }

        // Foreign key and relationship
        [Required]
        public int CustomerID { get; set; }
        public Customer Customer { get; set; }

        // Navigation property
        public ICollection<OrderItem> OrderItems { get; set; } = new List<OrderItem>();
    }
}