# backend/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "backend, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/OrderItems": {
      "get": {
        "tags": [
          "OrderItems"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/OrderItemReadDto"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "OrderItems"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/OrderItemCreateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/OrderItem"
            }
          }
        }
      }
    },
    "/api/OrderItems/{id}": {
      "get": {
        "tags": [
          "OrderItems"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/OrderItemReadDto"
            }
          }
        }
      },
      "put": {
        "tags": [
          "OrderItems"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/OrderUpdateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "OrderItems"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Orders/{id}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Orders"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/OrderUpdateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Orders/customer/{customerId}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "customerId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/OrderReadDto"
              }
            }
          }
        }
      }
    },
    "/api/Orders/paypal": {
      "post": {
        "tags": [
          "Orders"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/OrderCreateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          }
        }
      }
    },
    "/api/Orders/vnpay": {
      "post": {
        "tags": [
          "Orders"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/VnPaymentRequestModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Orders/paymentReturn": {
      "get": {
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
    "CartItem": {
      "type": "object",
      "properties": {
        "cartItemID": {
          "format": "int32",
          "type": "integer"
        },
        "customerID": {
          "format": "int32",
          "type": "integer"
        },
        "customer": {
          "$ref": "#/definitions/Customer"
        },
        "productSizeID": {
          "format": "int32",
          "type": "integer"
        },
        "productSize": {
          "$ref": "#/definitions/ProductSize"
        },
        "quantity": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "Customer": {
      "required": [
        "email",
        "firstName",
        "hashPassword",
        "lastName"
      ],
      "type": "object",
      "properties": {
        "customerID": {
          "format": "int32",
          "type": "integer"
        },
        "firstName": {
          "maxLength": 50,
          "minLength": 1,
          "type": "string"
        },
        "lastName": {
          "maxLength": 50,
          "minLength": 1,
          "type": "string"
        },
        "email": {
          "format": "email",
          "minLength": 1,
          "type": "string"
        },
        "emailConfirmed": {
          "type": "boolean"
        },
        "emailVerificationToken": {
          "maxLength": 200,
          "type": "string"
        },
        "phone": {
          "format": "tel",
          "type": "string"
        },
        "hashPassword": {
          "maxLength": 100,
          "minLength": 1,
          "type": "string"
        },
        "address": {
          "maxLength": 200,
          "type": "string"
        },
        "city": {
          "maxLength": 50,
          "type": "string"
        },
        "postalCode": {
          "maxLength": 20,
          "type": "string"
        },
        "orders": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Order"
          }
        },
        "cartItems": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CartItem"
          }
        },
        "totalAmount": {
          "format": "double",
          "type": "number",
          "readOnly": true
        }
      }
    },
...
    "Order": {
      "required": [
        "customerID",
        "dateTime",
        "status",
        "totalPrice",
        "transactionID"
      ],
      "type": "object",
      "properties": {
        "orderID": {
          "format": "int32",
          "type": "integer"
        },
        "dateTime": {
          "format": "date-time",
          "type": "string"
        },
        "totalPrice": {
          "format": "double",
          "type": "number"
        },
        "status": {
          "$ref": "#/definitions/OrderStatus"
        },
        "transactionID": {
          "minLength": 1,
          "type": "string"
        },
        "customerID": {
          "format": "int32",
          "type": "integer"
        },
        "customer": {
          "$ref": "#/definitions/Customer"
        },
        "orderItems": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderItem"
          }
        }
      }
    },
    "OrderCreateDto": {
      "type": "object",
      "properties": {
        "customerID": {
          "format": "int32",
          "type": "integer"
        },
        "transactionID": {
          "type": "string"
        },
        "totalPrice": {
          "format": "double",
          "type": "number"
        }
      }
    },
    "OrderItem": {
      "required": [
        "orderID",
        "productSizeID",
        "quantity"
      ],
      "type": "object",
      "properties": {
        "orderItemID": {
          "format": "int32",
          "type": "integer"
        },
        "quantity": {
          "format": "int32",
          "type": "integer"
        },
        "orderID": {
          "format": "int32",
          "type": "integer"
        },
        "productSizeID": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "OrderItemCreateDto": {
      "type": "object",
      "properties": {
        "productID": {
          "format": "int32",
          "type": "integer"
        },
        "quantity": {
          "format": "int32",
          "type": "integer"
        },
        "productSizeID": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "OrderItemReadDto": {
      "type": "object",
      "properties": {
        "orderItemID": {
          "format": "int32",
          "type": "integer"
        },
        "product": {
          "$ref": "#/definitions/ProductReadDto"
        },
        "quantity": {
          "format": "int32",
          "type": "integer"
        },
        "productSize": {
          "$ref": "#/definitions/ProductSizeReadDto"
        }
      }
    },
    "OrderReadDto": {
      "type": "object",
      "properties": {
        "orderID": {
          "format": "int32",
          "type": "integer"
        },
        "dateTime": {
          "format": "date-time",
          "type": "string"
        },
        "totalPrice": {
          "format": "double",
          "type": "number"
        },
        "status": {
          "type": "string"
        },
        "transactionID": {
          "type": "string"
        },
        "customerID": {
          "format": "int32",
          "type": "integer"
        },
        "customerName": {
          "type": "string"
        },
        "orderItems": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderItemReadDto"
          }
        }
      }
    },
    "OrderStatus": {
      "format": "int32",
      "enum": [
        0,
        1,
        2,
        3,
        4
      ],
      "type": "integer"
    },
    "OrderUpdateDto": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        }
      }
    },
...
    "ProductReadDto": {
      "type": "object",
      "properties": {
        "productID": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "brand": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "imageURL": {
          "type": "string"
        },
        "nameAlias": {
          "type": "string"
        },
        "categoryID": {
          "format": "int32",
          "type": "integer"
        },
        "categoryName": {
          "type": "string"
        },
        "productSizes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProductSizeReadDto"
          }
        }
      }
    },
    "ProductSize": {
      "required": [
        "price",
        "productID",
        "quantity",
        "size"
      ],
      "type": "object",
      "properties": {
        "productSizeID": {
          "format": "int32",
          "type": "integer"
        },
        "size": {
          "format": "int32",
          "type": "integer"
        },
        "price": {
          "format": "double",
          "type": "number"
        },
        "quantity": {
          "format": "int32",
          "type": "integer"
        },
        "productID": {
          "format": "int32",
          "type": "integer"
        },
        "orderItems": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderItem"
          }
        }
      }
    },
...
    "ProductSizeReadDto": {
      "type": "object",
      "properties": {
        "productSizeID": {
          "format": "int32",
          "type": "integer"
        },
        "size": {
          "format": "int32",
          "type": "integer"
        },
        "price": {
          "format": "double",
          "type": "number"
        },
        "quantity": {
          "format": "int32",
          "type": "integer"
        },
        "productID": {
          "format": "int32",
          "type": "integer"
        },
        "productName": {
          "type": "string"
        }
      }
    },
...
    "VnPaymentRequestModel": {
      "type": "object",
      "properties": {
        "orderId": {
          "format": "int32",
          "type": "integer"
        },
        "fullName": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "amount": {
          "format": "double",
          "type": "number"
        },
        "createdDate": {
          "format": "date-time",
          "type": "string"
        },
        "order": {
          "$ref": "#/definitions/OrderCreateDto"
        }
      }
    }
  }
}